fix: board and pieces scale correctly
After a lot of fiddling, and a lot of help from Codeium, I was able to
build the layout for the piece cells correctly. The board background can
be used properly, and the pieces are placed on them properly. Also, the
pieces scale correctly dependent on the window aspect ratio. This will
most likely need to be revisited in the future, when more elements are
added to the App. For now, it works well enough that I can move on to
implementing the functions to get the game to now actually work.