In Progres: Piece and Board Interaction
so far, I can click on pieces and squares, and they will echo the index
of the square they are located. I can "drag" a piece around the board,
but it can't be dropped, and it constantly "teleports" back to it's
original square and where I am holding it. The log gets filled with
constant drag messages.