feat: chess is playable
can play chess. At the moment, promotions are automatically queens.
Needd to implement a way to choose which piece to promote to. Also, need
to figure out how to get rid of the latency when making a move. I
suspect this is because of the way I've set up the way moves are made.
Will need to look into this more. After solving these two issues I can
consider this section of the project done, and can move on to
implenting how the different draws are detected.