feat: a promotion has been implemented
when a pawn reaches a promotion square, a pop up is created the allows
the player to choose a piece to promote to, no more auto-queen. It's not
the prettiest implementation, but it works. Will need to work on
cleaning it up during the "Refactoring". I want to make sure my library
works as intended, before updating this program to use the latest
library. I plan to work on positioning the pop up promotion selection
menu, as well as the size of the pop up menu. After which, I will begin
to work on being able to drag the pieces.